Added refactored the methods in the Grass class and also added variables where possible

+++ src/Main/Organisms/Plants/Plant.java
@@ -5,7 +5,7 @@ import Main.Helper.Transform;
 import Main.Organisms.Organism;
 
 public abstract class Plant extends Organism {
-    float growthInterval, growthTimer = 0;
+    double growthInterval, growthTimer = 0;
 
     //Constructors
     public Plant(Transform transform, float health, DNA dna){
